// src/models/requests/NewApplication.ts
export interface NewApplication {
    applicationCopy: File[];
    passportCopy: File[];
    planeCopy: File[];
    ownDocsCopy: File[];
    powerOfAttorneyCopy: File[];
    constituentDocsCopy: File[];
    otherDocs: File[];
    reason: string;
    city: string;
    address: string;
    maxPower: string;
    powerLevel: string;
    paymentsOption: string;
    provider: string;
}